feat: minimal binary (#11)
* feat: simple command to start go-execution-evm as server

* chore: add build target for evm-middleware in Makefile

Introduce a new build target to compile the evm-middleware binary and update the clean target to remove the build directory. Also, document the build process in the README and update the .gitignore to exclude the build directory.

* chore: add go formatting to --> Formatting markdownlint

* fix: add missing  in engineURL parameter

* fix: use same genesis in binary as in docker

* chore: go mod tidy

* feat: use cobra, introduce run command

Implemented a new `run` command using Cobra to start the EVM middleware, simplifying initialization and configuration through CLI flags. Refactored the main file to delegate execution to the command structure for better modularity and maintainability.

* fix: address linter errors

* chore: improve default RPC URLs and validate genesis hash format

Updated the default ETH and Engine API URLs to use localhost for better clarity and security. Introduced a `PreRunE` validation step to ensure the genesis hash is in a valid hex format, improving error handling and configuration robustness. Removed redundant initialization of the `runCmd` in the root command file.

* fix: repair go.mod